refactor(workspaces): share citation UI across rich text
